requirements.txt
asgiref==3.7.2
Django==3.2.20
django-dotenv==1.4.2
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
importlib-metadata==6.7.0
Markdown==3.4.3
Pillow==9.5.0
psycopg2==2.9.6
PyJWT==2.7.0
python-dotenv==0.21.1
pytz==2023.3
sqlparse==0.4.4
typing-extensions==4.7.1
zipp==3.15.0
